Split LdapConfig into Client and Authenticator config
