+++ plugin/trino-password-authenticators/src/test/java/io/trino/plugin/password/ldap/TestLdapConfig.java
@@ -14,7 +14,6 @@
  */
 package io.trino.plugin.password.ldap;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import io.airlift.units.Duration;
 import org.testng.annotations.Test;
@@ -40,20 +39,14 @@ public class TestLdapConfig
     @Test
     public void testDefault()
     {
-        assertRecordedDefaults(recordDefaults(LdapConfig.class)
+        assertRecordedDefaults(recordDefaults(LdapClientConfig.class)
                 .setLdapUrl(null)
                 .setAllowInsecure(false)
                 .setKeystorePath(null)
                 .setKeystorePassword(null)
                 .setTrustStorePath(null)
                 .setTruststorePassword(null)
-                .setUserBindSearchPatterns(" : ")
-                .setUserBaseDistinguishedName(null)
-                .setGroupAuthorizationSearchPattern(null)
-                .setBindDistingushedName(null)
-                .setBindPassword(null)
                 .setIgnoreReferrals(false)
-                .setLdapCacheTtl(new Duration(1, TimeUnit.HOURS))
                 .setLdapConnectionTimeout(null)
                 .setLdapReadTimeout(null));
     }
@@ -72,31 +65,19 @@ public class TestLdapConfig
                 .put("ldap.ssl.keystore.password", "12345")
                 .put("ldap.ssl.truststore.path", trustStoreFile.toString())
                 .put("ldap.ssl.truststore.password", "54321")
-                .put("ldap.user-bind-pattern", "uid=${USER},ou=org,dc=test,dc=com:uid=${USER},ou=alt")
-                .put("ldap.user-base-dn", "dc=test,dc=com")
-                .put("ldap.group-auth-pattern", "&(objectClass=user)(memberOf=cn=group)(user=username)")
-                .put("ldap.bind-dn", "CN=User Name,OU=CITY_OU,OU=STATE_OU,DC=domain,DC=domain_root")
-                .put("ldap.bind-password", "password1234")
                 .put("ldap.ignore-referrals", "true")
-                .put("ldap.cache-ttl", "2m")
                 .put("ldap.timeout.connect", "3m")
                 .put("ldap.timeout.read", "4m")
                 .buildOrThrow();
 
-        LdapConfig expected = new LdapConfig()
+        LdapClientConfig expected = new LdapClientConfig()
                 .setLdapUrl("ldaps://localhost:636")
                 .setAllowInsecure(true)
                 .setKeystorePath(keyStoreFile.toFile())
                 .setKeystorePassword("12345")
                 .setTrustStorePath(trustStoreFile.toFile())
                 .setTruststorePassword("54321")
-                .setUserBindSearchPatterns(ImmutableList.of("uid=${USER},ou=org,dc=test,dc=com", "uid=${USER},ou=alt"))
-                .setUserBaseDistinguishedName("dc=test,dc=com")
-                .setGroupAuthorizationSearchPattern("&(objectClass=user)(memberOf=cn=group)(user=username)")
-                .setBindDistingushedName("CN=User Name,OU=CITY_OU,OU=STATE_OU,DC=domain,DC=domain_root")
-                .setBindPassword("password1234")
                 .setIgnoreReferrals(true)
-                .setLdapCacheTtl(new Duration(2, TimeUnit.MINUTES))
                 .setLdapConnectionTimeout(new Duration(3, TimeUnit.MINUTES))
                 .setLdapReadTimeout(new Duration(4, TimeUnit.MINUTES));
 
@@ -106,30 +87,24 @@ public class TestLdapConfig
     @Test
     public void testValidation()
     {
-        assertValidates(new LdapConfig()
-                .setLdapUrl("ldaps://localhost")
-                .setUserBindSearchPatterns("uid=${USER},ou=org,dc=test,dc=com")
-                .setUserBaseDistinguishedName("dc=test,dc=com")
-                .setGroupAuthorizationSearchPattern("&(objectClass=user)(memberOf=cn=group)(user=username)"));
+        assertValidates(new LdapClientConfig()
+                .setLdapUrl("ldaps://localhost"));
 
-        assertValidates(new LdapConfig()
+        assertValidates(new LdapClientConfig()
                 .setLdapUrl("ldap://localhost")
-                .setAllowInsecure(true)
-                .setUserBindSearchPatterns("uid=${USER},ou=org,dc=test,dc=com")
-                .setUserBaseDistinguishedName("dc=test,dc=com")
-                .setGroupAuthorizationSearchPattern("&(objectClass=user)(memberOf=cn=group)(user=username)"));
+                .setAllowInsecure(true));
 
         assertFailsValidation(
-                new LdapConfig()
+                new LdapClientConfig()
                         .setLdapUrl("ldap://")
                         .setAllowInsecure(false),
                 "urlConfigurationValid",
                 "Connecting to the LDAP server without SSL enabled requires `ldap.allow-insecure=true`",
                 AssertTrue.class);
 
-        assertFailsValidation(new LdapConfig().setLdapUrl("localhost"), "ldapUrl", "Invalid LDAP server URL. Expected ldap:// or ldaps://", Pattern.class);
-        assertFailsValidation(new LdapConfig().setLdapUrl("ldaps:/localhost"), "ldapUrl", "Invalid LDAP server URL. Expected ldap:// or ldaps://", Pattern.class);
+        assertFailsValidation(new LdapClientConfig().setLdapUrl("localhost"), "ldapUrl", "Invalid LDAP server URL. Expected ldap:// or ldaps://", Pattern.class);
+        assertFailsValidation(new LdapClientConfig().setLdapUrl("ldaps:/localhost"), "ldapUrl", "Invalid LDAP server URL. Expected ldap:// or ldaps://", Pattern.class);
 
-        assertFailsValidation(new LdapConfig(), "ldapUrl", "may not be null", NotNull.class);
+        assertFailsValidation(new LdapClientConfig(), "ldapUrl", "may not be null", NotNull.class);
     }
 }
